move blueprints code to blueprints pkg
