--- examples/experimental/dagster-blueprints/dagster_blueprints/__init__.py
@@ -1,5 +1,0 @@
-from dagster._core.libraries import DagsterLibraryRegistry
-
-from .version import __version__ as __version__
-
-DagsterLibraryRegistry.register("dagster-blueprints", __version__)
