--- examples/experimental/dagster-blueprints/dagster_blueprints/py.typed
@@ -1,1 +1,0 @@
-partial
