+++ examples/experimental/dagster-blueprints/dagster_blueprints_tests/yaml_files/dir_of_single_blueprints/single_blueprint2.yaml
