+++ examples/experimental/dagster-blueprints/dagster_blueprints_tests/yaml_files/list_of_blueprints.yaml
