+++ examples/experimental/dagster-blueprints/dagster_blueprints_tests/yaml_files/single_blueprint_with_type.yaml
