+++ examples/experimental/dagster-blueprints/tox.ini
@@ -8,6 +8,7 @@ install_command = uv pip install {opts} {packages}
 deps =
   -e ../../../python_modules/dagster[test]
   -e ../../../python_modules/dagster-pipes
+  -e ../../../python_modules/libraries/dagster-databricks
   -e .
 allowlist_externals =
   /bin/bash
